requirements.txt
requests
python-dotenv
boto3
sqlalchemy
psycopg2-binary
flask
pandas
numpy
scikit-learn
matplotlib
